refactor(all): slight refactor to have cleaner code
