+++ src/app/chores/error.tsx
@@ -2,13 +2,13 @@
 
 import { useEffect } from 'react';
 
-export default function Error({
+const Error = ({
   error,
   reset,
 }: {
   error: Error & { digest?: string };
   reset: () => void;
-}) {
+}) => {
   useEffect(() => {
     console.error(error);
   }, [error]);
@@ -26,4 +26,6 @@ export default function Error({
       </button>
     </div>
   );
-}
+};
+
+export default Error;
