+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import { UserProvider } from '@auth0/nextjs-auth0/client';
 import './globals.css';
-import Profile from './components/organisms/Profile';
-import Sidebar from './components/organisms/Sidebar';
+import { Profile } from '@app/components/organisms/Profile';
+import { NavBar } from '@app/components/organisms/NavBar';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -22,7 +22,7 @@ export default function RootLayout({
       <UserProvider>
         <body className={`${inter.className} flex flex-row h-screen`}>
           <Profile />
-          <Sidebar />
+          <NavBar />
           <main className="p-8 w-4/5">{children}</main>
         </body>
       </UserProvider>
